+++ src/graphics/texture.rs
@@ -6,7 +6,7 @@ use std::rc::Rc;
 use image;
 
 use error::{Result, TetraError};
-use graphics::opengl::GLTexture;
+use graphics::opengl::{GLTexture, TextureFormat};
 use graphics::{self, DrawParams, Drawable, Rectangle};
 use Context;
 
@@ -25,9 +25,19 @@ impl Texture {
         let image = image::open(path).map_err(TetraError::Image)?.to_rgba();
         let (width, height) = image.dimensions();
 
-        let texture = ctx.gl.new_texture(width as i32, height as i32);
-        ctx.gl
-            .set_texture_data(&texture, &image, 0, 0, width as i32, height as i32);
+        let texture = ctx
+            .gl
+            .new_texture(width as i32, height as i32, TextureFormat::Rgba);
+
+        ctx.gl.set_texture_data(
+            &texture,
+            &image,
+            0,
+            0,
+            width as i32,
+            height as i32,
+            TextureFormat::Rgba,
+        );
 
         Ok(Texture::from_handle(texture))
     }