+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Inspired by FNA, the `TETRA_OPENGL_FORCE_CORE_PROFILE` environment variable can now be set to force the application to run using the 3.2 core profile. This might end up getting removed in favour of a more robust solution later on, but it's handy for testing (e.g. Renderdoc requires the core profile to be enabled).
 
+### Bug Fixes
+
+* The internal framebuffer is now an RGB texture instead of an RGBA texture - this was causing some strange issues with blending.
+
 ## 0.1.5 (December 8, 2018)
 
 ### Bug Fixes